tests: stabilize
Window when shown wasn't honouring its preferred height and was
slightly bigger. Seems window manager specific since with offscreen
QPA it works, only with X11 it doesn't. It's a bug either in Qt
or in the test code, the declarative widget layouting is fine and
passing all tests that are related to the grid layout.

Added XFAILs for QTBUG-66747